+++ tests/test_httpx_async.py
@@ -1,5 +1,8 @@
+from typing import Optional
+
 import pytest
 import httpx
+from httpx import content_streams
 
 from pytest_httpx import httpx_mock, HTTPXMock
 
@@ -383,3 +386,37 @@ async def test_httpx_mock_requests_json_body(httpx_mock: HTTPXMock):
 
         response = await client.put("http://test_url")
         assert response.json() == "string value"
+
+
+@pytest.mark.asyncio
+async def test_callback_raising_exception(httpx_mock: HTTPXMock):
+    def raise_timeout(
+        request: httpx.Request, timeout: Optional[httpx.Timeout]
+    ) -> httpx.Response:
+        raise httpx.exceptions.TimeoutException()
+
+    httpx_mock.add_callback(raise_timeout, "http://test_url")
+
+    async with httpx.AsyncClient() as client:
+        with pytest.raises(httpx.exceptions.TimeoutException):
+            await client.get("http://test_url")
+
+
+@pytest.mark.asyncio
+async def test_callback_returning_response(httpx_mock: HTTPXMock):
+    def custom_response(
+        request: httpx.Request, timeout: Optional[httpx.Timeout]
+    ) -> httpx.Response:
+        return httpx.Response(
+            status_code=200,
+            http_version="HTTP/1.1",
+            headers=[],
+            stream=content_streams.JSONStream({"url": str(request.url)}),
+            request=request,
+        )
+
+    httpx_mock.add_callback(custom_response, "http://test_url")
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("http://test_url")
+        assert response.json() == {"url": "http://test_url"}