+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mock async httpx requests as well.
 - Allow to provide files and boundary for multipart response.
 - Allow to provide data as a dictionary for multipart response.
+- Allow to provide callbacks that are executed upon reception of a request.
 
 ### Changed
 - method can now be provided even if not entirely upper cased.